BitGo Wallet Recovery Tool
Summary:
The BitGo Offline Recovery Tool provides BitGo customers with a mechanism to retrieve their Bitcoin stored with BitGo wallets agnostic of the status of BitGo.

This tool intentionally has no dependencies on the BitGo platform API.  It uses an external API (Chain) for retrieving blockchain data. This is in order to demonstrate that Bitcoin stored in wallets created through the BitGo service can be recovered independently of BitGo, using only the KeyCard provided when creating a wallet on BitGo along with the wallet password.

Reviewers: ryan, ben, benchan

Reviewed By: ben, benchan

Differential Revision: https://phabricator.bitgo.com/D1157