change how the tool discovers base addresses and searchs through child-key addresses to now look for activity on an address, rather than just balance
Summary:
We had a complaint from a user that the recovery tool was not working
for them. It turns out the tool won't work if you have a lot of activity on your
wallet and the first 20 addresses have all had their unspents used. Now the
child-key traversal logic is more robust and will search until it finds 20
sequentially-derived addresses which have never had any activity on them

Reviewers: mason, ben

Reviewed By: ben

Differential Revision: https://phabricator.bitgo.com/D2938